verson 0.3, can archive multiple files but not extract to origiginal directory. termination of project planned, will move from 7zip to rar

+++ test_archive.py
@@ -8,6 +8,36 @@ opj = os.path.join
 
 class TestArchive(unittest.TestCase):
 
+
+  def testMakeArchive2Files(self):
+    # Part A creates a single file, archives it, then deletes it
+    # Part B extracts the file again from the archive and checks the content
+    
+    # Part A:
+    file_helper.create_file(opj(self.tempdir,"testfile1"),"content-1")
+    file_helper.create_file(opj(self.tempdir,"testfile2"),"content-2")
+    
+    a = archive.Archive(opj(self.tempdir,"testarchive1.7zip"))
+    a.add_file(opj(self.tempdir,"testfile1"))
+    a.add_file(opj(self.tempdir,"testfile2"))
+
+    self.assertEqual(a.run(),0) 
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"testarchive1.7zip")),True)
+
+    file_helper.remove_file(opj(self.tempdir,"testfile1"))
+    file_helper.remove_file(opj(self.tempdir,"testfile2"))
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"testfile1")),False)
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"testfile2")),False)
+       
+    # Part B:
+    b = archive.Archive(opj(self.tempdir,"testarchive1.7zip"))
+    syscode = b.extract(target_path=self.tempdir)
+    self.assertEqual(syscode,0)
+
+    self.assertEqual(file_helper.file_head(opj(self.tempdir,"testfile1")),"content-1")
+    self.assertEqual(file_helper.file_head(opj(self.tempdir,"testfile2")),"content-2")
+
+
   def testMakeArchive(self):
     # Part A creates a single file, archives it, then deletes it
     # Part B extracts the file again from the archive and checks the content
@@ -59,9 +89,83 @@ class TestArchive(unittest.TestCase):
     a.add_file(opj(self.tempdir,"this-file-does-not-exist"))
     syscode = a.run() 
     self.assertNotEqual(syscode,0)
-        
+
+
+  def xtestMakeArchiveSubdir1(self):
+  
+    self.keep = True
+  
+    a = archive.Archive(opj(self.tempdir,"testarchive1.7zip"))
+
+    file_helper.mkdir_p(opj(self.tempdir,    "A","1","a"))
+    file_helper.create_file(opj(self.tempdir,"A","1","a","testfile-1"),"content-1")
+
+    a.add_dir(opj(self.tempdir              ,"A"))
+
+    self.assertEqual(a.run(),0) 
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"testarchive1.7zip")),True)
+
+    file_helper.remove_file(opj(self.tempdir,"A","1","a","testfile-1"))
+
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"A","1","a","testfile-1")),False)
+                                                                        
+    # test content of archive by extracting:                                                     
+    b = archive.Archive(opj(self.tempdir,"testarchive1.7zip"))          
+    syscode = b.extract(target_path=self.tempdir)
+    self.assertEqual(syscode,0)
+
+    print "TEMPDIR",self.tempdir
+
+    self.assertEqual(file_helper.file_head(opj(self.tempdir,"A","1","a","testfile-1")),"content-1")
+
+
+
+  def xtestMakeArchive2Trees(self):
+    # Part A creates a single file, archives it, then deletes it
+    # Part B extracts the file again from the archive and checks the content
+    
+    # Part A:
+    a = archive.Archive(opj(self.tempdir,"testarchive1.7zip"))
+
+    file_helper.mkdir_p(opj(self.tempdir,    "A","1","a"))
+    file_helper.create_file(opj(self.tempdir,"A","1","a","testfile-1"),"content-1")
+
+    a.add_dir(opj(self.tempdir              ,"A"))
+
+
+    file_helper.mkdir_p(opj(self.tempdir,    "B","1","a"))
+    file_helper.create_file(opj(self.tempdir,"B","1","a","testfile-2"),"content-2")
+
+    file_helper.mkdir_p(opj(self.tempdir,    "B","2","a"))
+    file_helper.create_file(opj(self.tempdir,"B","2","a","testfile-3"),"content-3")
+
+    a.add_dir(opj(self.tempdir,              "B"))
+
+    self.assertEqual(a.run(),0) 
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"testarchive1.7zip")),True)
+
+    file_helper.remove_file(opj(self.tempdir,"A","1","a","testfile-1"))
+    file_helper.remove_file(opj(self.tempdir,"B","1","a","testfile-2"))
+    file_helper.remove_file(opj(self.tempdir,"B","2","a","testfile-3"))
+
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"A","1","a","testfile-1")),False)
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"B","1","a","testfile-2")),False)
+    self.assertEqual(file_helper.file_exists(opj(self.tempdir,"B","2","a","testfile-3")),False)
+                                                                        
+    # Part B:                                                     
+    b = archive.Archive(opj(self.tempdir,"testarchive1.7zip"))          
+    syscode = b.extract(target_path=self.tempdir)
+    self.assertEqual(syscode,0)
+
+    print "TEMPDIR",self.tempdir
+
+    self.assertEqual(file_helper.file_head(opj(self.tempdir,"A","1","a","testfile-1")),"content-1")
+    self.assertEqual(file_helper.file_head(opj(self.tempdir,"B","1","a","testfile-2")),"content-2")
+    self.assertEqual(file_helper.file_head(opj(self.tempdir,"B","2","a","testfile-3")),"content-3")
+
     
   def setUp(self):
+    self.keep = False # set this to true to keep a resulting temp dir after testing
     #get a temp dir:
     self.tempdir = file_helper.get_random_temp_dir_name()
     #create the temp dir:
@@ -70,8 +174,14 @@ class TestArchive(unittest.TestCase):
     
   
   def tearDown(self):
+    if self.keep:
+      print "RESULT WAS KEPT: ================================="
+      print
+      print "cd", self.tempdir
+      print
+      print
+      return
     file_helper.destroy_dir_recursive(self.tempdir)
-
